Updated NCC GPU kernel
Successfully tested the kernel up to 11x11 image with a kernel size of 5x5, overlap = 0.5, and shift of 3x3. Resolution of disp image was 6x11. (Assuming the size of the displacement image will be pre-calculated in MATLAB). Right now each cuda block corresponds to one displacement point with thread number equal to shiftX*shiftY.